+++ tests/tests_pytorch/callbacks/test_lambda_function.py
@@ -48,10 +48,7 @@ def test_lambda_call(tmpdir):
         limit_val_batches=1,
         callbacks=[LambdaCallback(**hooks_args)],
     )
-    with pytest.deprecated_call(
-        match="`on_init_start` callback hook was deprecated in v1.6 and will be removed in v1.8."
-    ):
-        trainer.fit(model)
+    trainer.fit(model)
     ckpt_path = trainer.checkpoint_callback.best_model_path
 
     # raises KeyboardInterrupt and loads from checkpoint
@@ -64,17 +61,8 @@ def test_lambda_call(tmpdir):
         limit_predict_batches=1,
         callbacks=[LambdaCallback(**hooks_args)],
     )
-    with pytest.deprecated_call(
-        match="`on_init_start` callback hook was deprecated in v1.6 and will be removed in v1.8."
-    ):
-        trainer.fit(model, ckpt_path=ckpt_path)
-    with pytest.deprecated_call(
-        match="`on_init_start` callback hook was deprecated in v1.6 and will be removed in v1.8."
-    ):
-        trainer.test(model)
-    with pytest.deprecated_call(
-        match="`on_init_start` callback hook was deprecated in v1.6 and will be removed in v1.8."
-    ):
-        trainer.predict(model)
+    trainer.fit(model, ckpt_path=ckpt_path)
+    trainer.test(model)
+    trainer.predict(model)
 
     assert checker == hooks