+++ tests/tests_pytorch/models/test_hooks.py
@@ -483,10 +483,6 @@ def test_trainer_model_hook_system_fit(tmpdir, kwargs, automatic_optimization):
         track_grad_norm=1,
         **kwargs,
     )
-    assert called == [
-        dict(name="Callback.on_init_start", args=(trainer,)),
-        dict(name="Callback.on_init_end", args=(trainer,)),
-    ]
     trainer.fit(model)
     saved_ckpt = {
         "callbacks": ANY,
@@ -502,8 +498,6 @@ def test_trainer_model_hook_system_fit(tmpdir, kwargs, automatic_optimization):
         saved_ckpt[trainer.precision_plugin.__class__.__qualname__] = ANY
     device = torch.device("cuda:0" if "accelerator" in kwargs and kwargs["accelerator"] == "gpu" else "cpu")
     expected = [
-        dict(name="Callback.on_init_start", args=(trainer,)),
-        dict(name="Callback.on_init_end", args=(trainer,)),
         dict(name="configure_callbacks"),
         dict(name="prepare_data"),
         dict(name="Callback.on_before_accelerator_backend_setup", args=(trainer, model)),
@@ -600,10 +594,6 @@ def test_trainer_model_hook_system_fit_no_val_and_resume_max_epochs(tmpdir):
         callbacks=[callback],
         track_grad_norm=1,
     )
-    assert called == [
-        dict(name="Callback.on_init_start", args=(trainer,)),
-        dict(name="Callback.on_init_end", args=(trainer,)),
-    ]
 
     # resume from checkpoint with HookedModel
     model = HookedModel(called)
@@ -620,8 +610,6 @@ def test_trainer_model_hook_system_fit_no_val_and_resume_max_epochs(tmpdir):
     }
     saved_ckpt = {**loaded_ckpt, "global_step": 4, "epoch": 1}
     expected = [
-        dict(name="Callback.on_init_start", args=(trainer,)),
-        dict(name="Callback.on_init_end", args=(trainer,)),
         dict(name="configure_callbacks"),
         dict(name="prepare_data"),
         dict(name="Callback.on_before_accelerator_backend_setup", args=(trainer, model)),
@@ -697,10 +685,6 @@ def test_trainer_model_hook_system_fit_no_val_and_resume_max_steps(tmpdir):
         callbacks=[callback],
         track_grad_norm=1,
     )
-    assert called == [
-        dict(name="Callback.on_init_start", args=(trainer,)),
-        dict(name="Callback.on_init_end", args=(trainer,)),
-    ]
 
     trainer.fit(model, ckpt_path=best_model_path)
     loaded_ckpt = {
@@ -715,8 +699,6 @@ def test_trainer_model_hook_system_fit_no_val_and_resume_max_steps(tmpdir):
     }
     saved_ckpt = {**loaded_ckpt, "global_step": steps_after_reload}
     expected = [
-        dict(name="Callback.on_init_start", args=(trainer,)),
-        dict(name="Callback.on_init_end", args=(trainer,)),
         dict(name="configure_callbacks"),
         dict(name="prepare_data"),
         dict(name="Callback.on_before_accelerator_backend_setup", args=(trainer, model)),
@@ -778,10 +760,6 @@ def test_trainer_model_hook_system_eval(tmpdir, batches, verb, noun, dataloader,
         enable_model_summary=False,
         callbacks=[callback],
     )
-    assert called == [
-        dict(name="Callback.on_init_start", args=(trainer,)),
-        dict(name="Callback.on_init_end", args=(trainer,)),
-    ]
     fn = getattr(trainer, verb)
     fn(model, verbose=False)
     hooks = [
@@ -798,8 +776,6 @@ def test_trainer_model_hook_system_eval(tmpdir, batches, verb, noun, dataloader,
         dict(name=f"on_{noun}_model_train"),
     ]
     expected = [
-        dict(name="Callback.on_init_start", args=(trainer,)),
-        dict(name="Callback.on_init_end", args=(trainer,)),
         dict(name="configure_callbacks"),
         dict(name="prepare_data"),
         dict(name="Callback.on_before_accelerator_backend_setup", args=(trainer, model)),
@@ -822,14 +798,8 @@ def test_trainer_model_hook_system_predict(tmpdir):
     trainer = Trainer(
         default_root_dir=tmpdir, limit_predict_batches=batches, enable_progress_bar=False, callbacks=[callback]
     )
-    assert called == [
-        dict(name="Callback.on_init_start", args=(trainer,)),
-        dict(name="Callback.on_init_end", args=(trainer,)),
-    ]
     trainer.predict(model)
     expected = [
-        dict(name="Callback.on_init_start", args=(trainer,)),
-        dict(name="Callback.on_init_end", args=(trainer,)),
         dict(name="configure_callbacks"),
         dict(name="prepare_data"),
         dict(name="Callback.on_before_accelerator_backend_setup", args=(trainer, model)),