+++ src/pytorch_lightning/CHANGELOG.md
@@ -221,6 +221,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - Removed the deprecated `Trainer.use_amp` and `LightningModule.use_amp` attributes ([#14832](https://github.com/Lightning-AI/lightning/pull/14832))
 
+- Removed deprecated callback hooks
+  * `Callback.on_init_start`
+  * `Callback.on_init_end`
+
 
 ### Fixed